Break up typescript HMR tests
