+++ test/integration/typescript-error-hmr/next.config.js
