+++ test/integration/typescript-error-hmr/pages/type-error-recover.tsx
