+++ test/integration/typescript-hmr/pages/hello.tsx
